dev-requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.12.0
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-js-asset==2.0.0
django-mptt==0.14.0
django-pytest==0.2.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
